Fold the write implementation into one class
Content_Write and Content_Write_Rest read as the native/REST pair the read
ability has, and there is no such pair here: the endpoint is the only way
these abilities write. One class says that without a docblock having to.
