+++ includes/Abilities/Content/Content_Write.php
@@ -14,6 +14,7 @@ namespace WordPress\AI\Abilities\Content;
 use WP_Error;
 use WP_Post;
 use WP_Post_Type;
+use WordPress\AI\Abilities\Rest\Rest_Backend;
 
 // Exit if accessed directly.
 defined( 'ABSPATH' ) || exit;
@@ -24,17 +25,21 @@ defined( 'ABSPATH' ) || exit;
  * Registers `core/content-create`, `core/content-update` and `core/content-delete`, the
  * write counterparts to `core/content-query`.
  *
- * These abilities only have a REST-backed implementation: {@see Content_Write_Rest} calls
- * the posts endpoint of the post type, so the capability checks, sanitization, slug
- * handling, revisions and hooks are the ones the controller already performs. The
- * permission callbacks here are the abilities' own gate in front of that, and they answer
- * a narrower question than REST does: the post type has to be exposed to abilities.
+ * There is one implementation, and it is the REST API: every write is a request to the
+ * posts endpoint of the post type, so the capability checks, sanitization, slug handling,
+ * revisions and hooks are the ones the controller already performs. Nothing here reaches
+ * `wp_insert_post()` or its siblings directly.
+ *
+ * The permission callbacks are the abilities' own gate in front of that, and they answer a
+ * narrower question than REST does: the post type has to be exposed to abilities.
  *
  * @internal This class should not be used outside the plugin and there is no guarantee of backwards compatibility.
  *
  * @since x.x.x
  */
 final class Content_Write {
+
+	use Post_Type_Route;
 
 	/**
 	 * The ability category the write abilities belong to.
@@ -69,21 +74,12 @@ final class Content_Write {
 	private Content $content;
 
 	/**
-	 * The REST-backed implementation.
-	 *
-	 * @since x.x.x
-	 * @var \WordPress\AI\Abilities\Content\Content_Write_Rest
-	 */
-	private Content_Write_Rest $rest;
-
-	/**
 	 * Constructor.
 	 *
 	 * @since x.x.x
 	 */
 	public function __construct() {
 		$this->content = new Content();
-		$this->rest    = new Content_Write_Rest();
 	}
 
 	/**
@@ -202,7 +198,7 @@ final class Content_Write {
 			return $this->not_found_error();
 		}
 
-		return $this->rest->create_post( $exposed[ $post_type ], $input, $this->normalize_fields( $input ) );
+		return $this->create_post( $exposed[ $post_type ], $input, $this->normalize_fields( $input ) );
 	}
 
 	/**
@@ -221,7 +217,7 @@ final class Content_Write {
 			return $this->not_found_error();
 		}
 
-		return $this->rest->update_post( $post, $input, $this->normalize_fields( $input ) );
+		return $this->update_post( $post, $input, $this->normalize_fields( $input ) );
 	}
 
 	/**
@@ -242,7 +238,7 @@ final class Content_Write {
 
 		$force = ! empty( $input['force'] );
 
-		return $this->rest->delete_post( $post, $force, $this->normalize_fields( $input ) );
+		return $this->delete_post( $post, $force, $this->normalize_fields( $input ) );
 	}
 
 	/**
@@ -604,6 +600,238 @@ final class Content_Write {
 				'post'    => $this->get_post_output_schema(),
 			),
 		);
+	}
+
+	/**
+	 * Ability input fields that map to a REST parameter of the same name.
+	 *
+	 * The read abilities flatten the REST sub-objects, so they report `title_raw`. A write
+	 * sets the raw value, which the endpoint takes as plain `title`.
+	 *
+	 * @since x.x.x
+	 * @var list<string>
+	 */
+	private const WRITABLE_FIELDS = array( // phpcs:ignore SlevomatCodingStandard.Classes.DisallowMultiConstantDefinition -- This is used as an array const.
+		'title',
+		'content',
+		'excerpt',
+		'status',
+		'slug',
+		'date',
+		'date_gmt',
+		'author',
+		'parent',
+		'password',
+		'menu_order',
+		'comment_status',
+		'ping_status',
+		'sticky',
+		'template',
+		'format',
+	);
+
+	/**
+	 * Creates a post through the REST API.
+	 *
+	 * @since x.x.x
+	 *
+	 * @param \WP_Post_Type        $post_type_object The post type to create in.
+	 * @param array<string, mixed> $input            The ability input.
+	 * @param list<string>         $fields           The fields to report on the created post.
+	 * @return array<string, mixed>|\stdClass|\WP_Error The created post, or a WP_Error on failure.
+	 */
+	private function create_post( WP_Post_Type $post_type_object, array $input, array $fields ) {
+		$response = $this->request(
+			$post_type_object,
+			static function ( string $route, array $params ) {
+				return Rest_Backend::post( $route, $params );
+			},
+			$this->route( $post_type_object ),
+			$this->to_rest_params( $input )
+		);
+
+		if ( is_wp_error( $response ) ) {
+			return $response;
+		}
+
+		return $this->read_back( $response, $fields );
+	}
+
+	/**
+	 * Updates a post through the REST API.
+	 *
+	 * Only the fields present in the input are sent, so an update leaves every field it
+	 * does not name alone rather than resetting it to a default.
+	 *
+	 * @since x.x.x
+	 *
+	 * @param \WP_Post             $post   The post to update.
+	 * @param array<string, mixed> $input  The ability input.
+	 * @param list<string>         $fields The fields to report on the updated post.
+	 * @return array<string, mixed>|\stdClass|\WP_Error The updated post, or a WP_Error on failure.
+	 */
+	private function update_post( WP_Post $post, array $input, array $fields ) {
+		$post_type_object = get_post_type_object( $post->post_type );
+		if ( ! $post_type_object instanceof WP_Post_Type ) {
+			return $this->not_found_error();
+		}
+
+		$response = $this->request(
+			$post_type_object,
+			static function ( string $route, array $params ) {
+				return Rest_Backend::post( $route, $params );
+			},
+			$this->route( $post_type_object ) . '/' . (int) $post->ID,
+			$this->to_rest_params( $input )
+		);
+
+		if ( is_wp_error( $response ) ) {
+			return $response;
+		}
+
+		return $this->read_back( $response, $fields );
+	}
+
+	/**
+	 * Deletes a post through the REST API.
+	 *
+	 * The endpoint answers the two cases differently: forcing returns a `deleted` envelope
+	 * carrying the post as it was, while trashing returns the trashed post itself. Both are
+	 * reported the same way here, with `deleted` telling them apart.
+	 *
+	 * @since x.x.x
+	 *
+	 * @param \WP_Post     $post   The post to delete.
+	 * @param bool         $force  Whether to delete permanently instead of trashing.
+	 * @param list<string> $fields The fields to report on the deleted post.
+	 * @return array{deleted: bool, post: array<string, mixed>|\stdClass}|\WP_Error The outcome, or a WP_Error on failure.
+	 */
+	private function delete_post( WP_Post $post, bool $force, array $fields ) {
+		$post_type_object = get_post_type_object( $post->post_type );
+		if ( ! $post_type_object instanceof WP_Post_Type ) {
+			return $this->not_found_error();
+		}
+
+		/*
+		 * A forced delete leaves nothing to read afterwards, so the post is read while it
+		 * is still there. Trashing keeps the row, and it is read back after the request so
+		 * the reported status is the one the delete left behind.
+		 */
+		$previous = $force ? ( new Content_Rest() )->get_post( $post, $fields ) : null;
+		if ( is_wp_error( $previous ) ) {
+			return $previous;
+		}
+
+		$response = $this->request(
+			$post_type_object,
+			static function ( string $route, array $params ) {
+				return Rest_Backend::delete( $route, $params );
+			},
+			$this->route( $post_type_object ) . '/' . (int) $post->ID,
+			$force ? array( 'force' => true ) : array()
+		);
+
+		if ( is_wp_error( $response ) ) {
+			return $response;
+		}
+
+		if ( null !== $previous ) {
+			return array(
+				'deleted' => true,
+				'post'    => $previous,
+			);
+		}
+
+		$trashed = $this->read_back( $response, $fields );
+		if ( is_wp_error( $trashed ) ) {
+			return $trashed;
+		}
+
+		return array(
+			'deleted' => false,
+			'post'    => $trashed,
+		);
+	}
+
+	/**
+	 * Runs a request against a post type's posts endpoint and returns its data.
+	 *
+	 * Wraps the call in the post type exposure and global post context handling every
+	 * content request needs, so a failure still restores both.
+	 *
+	 * @since x.x.x
+	 *
+	 * @param \WP_Post_Type        $post_type_object The post type being addressed.
+	 * @param callable             $dispatch         Receives the route and parameters, returns the response.
+	 * @param string               $route            The REST route.
+	 * @param array<string, mixed> $params           Request parameters.
+	 * @return array<mixed>|\WP_Error The response data, or the error the endpoint returned.
+	 */
+	private function request( WP_Post_Type $post_type_object, callable $dispatch, string $route, array $params ) {
+		$restore_post_type = $this->prepare_post_type( $post_type_object );
+		$restore_context   = $this->capture_post_context();
+
+		try {
+			$response = $dispatch( $route, $params );
+		} finally {
+			$restore_context();
+			$restore_post_type();
+		}
+
+		if ( is_wp_error( $response ) ) {
+			return $response;
+		}
+
+		return Rest_Backend::data( $response );
+	}
+
+	/**
+	 * Reports a written post in the shape the read ability returns.
+	 *
+	 * The write response already carries the post, but reading it back through
+	 * {@see Content_Rest} keeps one field mapping instead of two that can drift.
+	 *
+	 * @since x.x.x
+	 *
+	 * @param array<mixed> $data   The write response data.
+	 * @param list<string> $fields The fields to report.
+	 * @return array<string, mixed>|\stdClass|\WP_Error The post data, or a WP_Error on failure.
+	 */
+	private function read_back( array $data, array $fields ) {
+		$post = isset( $data['id'] ) ? get_post( (int) $data['id'] ) : null;
+
+		// The endpoint answers a write with the written post, so a response without a
+		// readable ID is one the mapping cannot read.
+		if ( ! $post instanceof WP_Post ) {
+			return Rest_Backend::unexpected_response_error();
+		}
+
+		return ( new Content_Rest() )->get_post( $post, $fields );
+	}
+
+	/**
+	 * Maps the ability input to REST request parameters.
+	 *
+	 * Fields absent from the input stay absent, which is what makes a partial update
+	 * partial. `post_type` is not sent: it is carried by the route.
+	 *
+	 * @since x.x.x
+	 *
+	 * @param array<string, mixed> $input The ability input.
+	 * @return array<string, mixed> The REST request parameters.
+	 */
+	private function to_rest_params( array $input ): array {
+		$params = array();
+
+		foreach ( self::WRITABLE_FIELDS as $field ) {
+			if ( ! array_key_exists( $field, $input ) ) {
+				continue;
+			}
+
+			$params[ $field ] = $input[ $field ];
+		}
+
+		return $params;
 	}
 
 	/**
